Completed migration to KNIME 3.0 compatibility

+++ com.vernalis.knime.flowcontrol/src/com/vernalis/knime/flowcontrol/nodes/timedloops/abstrct/loopstart/TimedLoopStart.java
@@ -63,6 +63,6 @@ public interface TimedLoopStart {
 	 * 
 	 * @return The iteration counter
 	 */
-	public int getIteration();
+	public long getIteration();
 
 }
